[feat] : 뉴스 자세히 보기 화면 인자 넘겨서 출력
