+++ app/src/main/res/layout/fragment_detail_news.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
         <import type="com.wanted.preob.newsapp.domain.model.News" />
@@ -17,6 +18,92 @@
         <include
             android:id="@+id/detail_news_header"
             layout="@layout/layout_header" />
+
+        <androidx.core.widget.NestedScrollView
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            app:layout_constraintVertical_weight="1"
+            android:paddingHorizontal="@dimen/bg_margin_default"
+            android:layout_marginTop="20dp"
+            app:layout_constraintTop_toBottomOf="@id/detail_news_header">
+
+            <androidx.constraintlayout.widget.ConstraintLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content">
+
+                <TextView
+                    android:id="@+id/detail_news_title_tv"
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:text="@{news.title}"
+                    style="@style/T1"
+                    android:textStyle="bold"
+                    android:textColor="@color/title_color"
+                    app:layout_constraintTop_toTopOf="parent"
+                    app:layout_constraintEnd_toStartOf="@id/detail_news_saved_btn"
+                    app:layout_constraintStart_toStartOf="parent"/>
+
+
+                <ImageView
+                    android:id="@+id/detail_news_saved_btn"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:src="@{news.isSaved ? @drawable/ic_star_active : @drawable/ic_star_inactive}"
+                    app:layout_constraintTop_toTopOf="@id/detail_news_title_tv"
+                    app:layout_constraintEnd_toEndOf="parent" />
+
+                <TextView
+                    android:id="@+id/detail_news_author_tv"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="@{news.author}"
+                    style="@style/Sub1"
+                    android:textStyle="bold"
+                    android:textColor="@color/sub_color"
+                    android:layout_marginTop="14dp"
+                    app:layout_constraintTop_toBottomOf="@id/detail_news_title_tv"
+                    app:layout_constraintStart_toStartOf="parent" />
+
+                <TextView
+                    android:id="@+id/detail_news_time_tv"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="@{news.time}"
+                    style="@style/Sub1"
+                    android:textStyle="bold"
+                    android:textColor="@color/gray"
+                    android:layout_marginStart="10dp"
+                    app:layout_constraintTop_toTopOf="@id/detail_news_author_tv"
+                    app:layout_constraintBottom_toBottomOf="@id/detail_news_author_tv"
+                    app:layout_constraintStart_toEndOf="@id/detail_news_author_tv"/>
+
+                <androidx.cardview.widget.CardView
+                    android:id="@+id/detail_news_image_cv"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_marginTop="20dp"
+                    app:cardCornerRadius="10dp"
+                    app:layout_constraintTop_toBottomOf="@id/detail_news_author_tv">
+
+                    <ImageView
+                        android:id="@+id/detail_news_iv"
+                        bindImage="@{news.imgUrl}"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content" />
+                </androidx.cardview.widget.CardView>
+
+                <TextView
+                    android:id="@+id/detail_news_content_tv"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:text="@{news.content}"
+                    style="@style/Sub2"
+                    android:textStyle="bold"
+                    android:textColor="@color/black"
+                    android:layout_marginTop="@dimen/bg_margin_default"
+                    app:layout_constraintTop_toBottomOf="@id/detail_news_image_cv" />
+            </androidx.constraintlayout.widget.ConstraintLayout>
+        </androidx.core.widget.NestedScrollView>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
